refactor(texteditor): remove duplicate logic from math/image composables
