+++ contentcuration/contentcuration/frontend/shared/views/TipTapEditor/TipTapEditor/composables/useImageHandling.js
@@ -1,90 +1,42 @@
-import { ref, onMounted, onUnmounted, watch } from 'vue';
+import { ref, onMounted, onUnmounted } from 'vue';
+import { useModalPositioning } from './useModalPositioning';
 
 export function useImageHandling(editor) {
   const modalMode = ref(null);
   const modalInitialData = ref({});
-  const popoverStyle = ref({});
   const editingNodePos = ref(null);
-  const isModalCentered = ref(false);
 
-  // to store the button element that opened the popover
-  const anchorElement = ref(null);
-
-  const updatePosition = () => {
-    // If there's no anchor (centered), do nothing.
-    if (!anchorElement.value || isModalCentered.value) {
-      return;
-    }
-    const rect = anchorElement.value.getBoundingClientRect();
-    popoverStyle.value = {
-      position: 'fixed',
-      top: `${rect.bottom + 5}px`,
-      left: `${rect.right}px`,
-      transform: 'translateX(-100%)',
-    };
-  };
-
-  const setCenteredPosition = () => {
-    isModalCentered.value = true;
-    anchorElement.value = null; // Ensure we don't try to update a centered modal
-    popoverStyle.value = {
-      position: 'fixed',
-      top: '50%',
-      left: '50%',
-      transform: 'translate(-50%, -50%)',
-    };
-  };
-
-  const openCreateModal = ({ file = null, targetElement = null } = {}) => {
-    if (targetElement) {
-      // Save the anchor element and run the initial position calculation
-      anchorElement.value = targetElement;
-      isModalCentered.value = false;
-      updatePosition();
-    } else {
-      setCenteredPosition();
-    }
-    modalInitialData.value = { file };
-    modalMode.value = 'create';
-  };
-
-  const openEditModal = ({ pos, attrs }) => {
-    setCenteredPosition();
-    editingNodePos.value = pos;
-    modalInitialData.value = { ...attrs };
-    modalMode.value = 'edit';
-  };
+  const {
+    isModalOpen,
+    popoverStyle,
+    isModalCentered,
+    openModal,
+    closeModal: closeModalBase,
+    setupClickOutside,
+    cleanup,
+  } = useModalPositioning();
 
   const closeModal = () => {
     modalMode.value = null;
     modalInitialData.value = {};
     editingNodePos.value = null;
-    isModalCentered.value = false;
-    anchorElement.value = null; // Clear the anchor on close
+    closeModalBase();
   };
 
-  // This watcher now also manages the crucial scroll/resize listeners.
-  watch(modalMode, newMode => {
-    const clickOutsideHandler = event => {
-      const popover = document.querySelector('.image-upload-modal');
-      if (popover && !popover.contains(event.target)) {
-        closeModal();
-      }
-    };
+  setupClickOutside('.image-upload-modal', closeModal);
 
-    if (newMode) {
-      setTimeout(() => {
-        // Use setTimeout to avoid capturing the opening click
-        document.addEventListener('mousedown', clickOutsideHandler, true);
-        window.addEventListener('scroll', updatePosition, true);
-        window.addEventListener('resize', updatePosition, true);
-      }, 0);
-    } else {
-      document.removeEventListener('mousedown', clickOutsideHandler, true);
-      window.removeEventListener('scroll', updatePosition, true);
-      window.removeEventListener('resize', updatePosition, true);
-    }
-  });
+  const openCreateModal = ({ file = null, targetElement = null } = {}) => {
+    modalInitialData.value = { file };
+    modalMode.value = 'create';
+    openModal({ targetElement });
+  };
+
+  const openEditModal = ({ pos, attrs }) => {
+    editingNodePos.value = pos;
+    modalInitialData.value = { ...attrs };
+    modalMode.value = 'edit';
+    openModal({ centered: true }); // Edit mode is always centered
+  };
 
   const handleInsert = async data => {
     if (!data.src || !editor?.value) return;
@@ -118,22 +70,6 @@ export function useImageHandling(editor) {
     closeModal();
   };
 
-  watch(modalMode, mode => {
-    const handler = event => {
-      const popover = document.querySelector('.image-upload-modal');
-      if (popover && !popover.contains(event.target)) {
-        closeModal();
-      }
-    };
-    if (mode) {
-      setTimeout(() => {
-        document.addEventListener('mousedown', handler, true);
-      }, 0);
-    } else {
-      document.removeEventListener('mousedown', handler, true);
-    }
-  });
-
   onMounted(() => {
     if (editor?.value) {
       editor.value.on('open-image-editor', openEditModal);
@@ -144,6 +80,7 @@ export function useImageHandling(editor) {
     if (editor?.value) {
       editor.value.off('open-image-editor', openEditModal);
     }
+    cleanup();
   });
 
   return {
@@ -151,6 +88,7 @@ export function useImageHandling(editor) {
     modalInitialData,
     popoverStyle,
     isModalCentered,
+    isModalOpen,
     openCreateModal,
     closeModal,
     handleInsert,
